+++ lbz/exceptions.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 from http import HTTPStatus
-from typing import TYPE_CHECKING, Optional, Union
+from typing import TYPE_CHECKING, Generator, Optional, Type, Union
 
 from lbz.response import Response
 
@@ -352,3 +352,12 @@ class NetworkAuthenticationRequired(LambdaFWServerException):
 
     message = HTTPStatus.NETWORK_AUTHENTICATION_REQUIRED.description
     status_code = HTTPStatus.NETWORK_AUTHENTICATION_REQUIRED.value
+
+
+def all_lbz_errors(
+    cls: Type[LambdaFWException] = LambdaFWException,
+) -> Generator[Type[LambdaFWException], None, None]:
+    for subcls in cls.__subclasses__():
+        if subcls not in [LambdaFWClientException, LambdaFWServerException]:
+            yield subcls
+        yield from all_lbz_errors(subcls)